+++ extensions/iceworks-material-helper/webpack.config.js
@@ -14,6 +14,7 @@ const config = {
   externals: {
     vscode: 'commonjs vscode',
     prettier: 'commonjs prettier',
+    '@iceworks/code-generator': 'commonjs @iceworks/code-generator',
   },
   resolve: {
     extensions: ['.ts', '.js'],